+++ include/Small/All.hpp
@@ -2,7 +2,7 @@
  * @file All.hpp
  * @author Alexandre Plateau (lexplt.dev@gmail.com)
  * @brief Include the needed files to start a project using the library
- * @version 0.2
+ * @version 0.3
  * @date 2020-04-05
  * 
  * @copyright Copyright (c) 2020